structures cache coverage float (todo: structures cache rebuild needed)

+++ HarmonyDB.Index/HarmonyDB.Index.Analysis/Models/Index/StructureLink.cs
@@ -1,3 +1,3 @@
 ﻿namespace HarmonyDB.Index.Analysis.Models.Index;
 
-public record StructureLink(string Normalized, string ExternalId, byte NormalizationRoot, short Occurrences, short Successions, float Coverage);
+public record StructureLink(string Normalized, string ExternalId, byte NormalizationRoot, float Occurrences, float Successions, float Coverage);
